python-task-2/requirements.txt
cffi==1.17.0
cryptography==43.0.0
networkx==3.3
psycopg2-binary==2.9.9
pycparser==2.22
PyMySQL==1.1.1
pymysql2==1.3.3
